Created order confirmation page component at menu page
