--- src/test/java/pageComponents/OrderSummaryPageComponent.java
@@ -1,4 +1,0 @@
-package pageComponents;
-
-public class OrderSummaryPageComponent {
-}
